more winegcc warnings.
git-svn-id: svn://fly.thruhere.net/forsaken@1399 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/d3d9/title.c
@@ -10610,7 +10610,7 @@ void InitAvgFrameRateGlobals( MENU *Menu )
 {
     sprintf( DemoAvgFpsText, "Avg %.2f FPS", DemoAvgFps );
 	sprintf( DemoTotalFramesText, "%d frames %d.%02d",
-		DemoGameLoops,
+		(int)DemoGameLoops,
 		(int) DemoTotalTime / 60, (int) DemoTotalTime % 60 );
 }
